Exported some art, I dunno if it's useful.
I made the UV of the rocket on the sheet from the bottom to the top, maybe we'll make it spawn with some kind of shader that display the texture? or we could have it scaled like balloons that inflates I dunno. I splitted stuff to make it more usable?